Move self02 main method into seperate module
